+++ src/Message.cc
@@ -1,6 +1,18 @@
 #include "Message.h"
 
-Message::Message() : messageType(MessageType::NONE) {
+#include "State.h"
+
+Message::Message() : messageType(Message::NONE) {
+
+}
+
+void Message::setNextState(State* next) {
+
+	if(next) {
+		nextState = next;
+	} else {
+		messageType = Message::QUIT;
+	}
 
 }